+++ Cargo.toml
@@ -10,15 +10,19 @@ num-bigint = "0.4.6"
 
 tokio = { version = "1.38.0", features = ["full"] }
 
-alloy = { version = "0.5.4", features = ["providers"] }
-alloy-sol-types = { version = "0.8.14" }
-alloy-primitives = { version = "0.8.9" }
+alloy = { version = "0.5.4", features = ["providers"], optional = true }
+alloy-sol-types = { version = "0.8.14", optional = true }
+alloy-primitives = { version = "0.8.9", optional = true }
 tycho-core = { git = "https://github.com/propeller-heads/tycho-indexer.git", package = "tycho-core", tag = "0.46.0" }
 hex = "0.4.3"
 anyhow = "1.0.95"
 num-traits = "0.2.19"
 serde = { version = "1.0.217", features = ["derive"] }
 serde_json = "1.0.135"
+
+[features]
+default = ["evm"]
+evm = ["alloy", "alloy-sol-types", "alloy-primitives"]
 
 [profile.bench]
 debug = true
