--- src/encoding/approvals/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod approvals_manager;
-pub(crate) mod interface;
-mod permit2;
