+++ src/encoding/evm/approvals/mod.rs
@@ -1,0 +1,2 @@
+pub mod permit2;
+pub mod protocol_approvals_manager;
