+++ src/encoding/evm/approvals/protocol_approvals_manager.rs
