+++ src/encoding/evm/mod.rs
@@ -1,0 +1,5 @@
+pub mod approvals;
+mod router_encoder;
+mod strategy_encoder;
+mod swap_encoder;
+mod utils;
