+++ src/encoding/mod.rs
@@ -1,7 +1,7 @@
-mod approvals;
+#[cfg(feature = "evm")]
+mod evm;
 mod models;
 mod router_encoder;
 mod strategy_encoder;
-mod strategy_selector;
 mod swap_encoder;
-mod utils;
+mod user_approvals_manager;
