+++ src/encoding/user_approvals_manager.rs
